requirements.txt
streamlit
requests
speechrecognition
pytesseract
transformers
torch
PyMuPDF
python-docx
Pillow
folium
streamlit-folium
